Added Search Autocomplete
Added autocomplete to the app. Results appear when the query is more
than 3 characters long. Tapping a result will move the map camera to
that result.

I tried to leave the implementation simple so we could adjust it after
some feedback.

I also removed some local files from git tracking. They were already in
the .gitignore, but may have been added to the project before the
.gitignore change. So, you may want to backup your local files.